requirements.txt
appnope==0.1.0
boto==2.38.0
click==6.2
dateparser==0.3.1
decorator==4.0.4
dj-database-url==0.3.0
Django==1.8.9
django-admin-sortable==2.0.3
django-appconf==1.0.1
django-imagekit==3.3
django-js-reverse==0.6.1
django-polymorphic==0.7.2
django-rq==0.8.0
django-ses-backend==0.1.1
django-simple-history==1.7.0
django-solo==1.1.1
django-versatileimagefield==1.3
et-xmlfile==1.0.1
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
jdatetime==1.7.1
jdcal==1.2
openpyxl==2.3.2
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
pilkit==1.1.13
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
python-dateutil==2.4.2
PyYAML==3.11
redis==2.10.5
reportlab==3.2.0
rq==0.5.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
validate-email==1.3
wheel==0.24.0
